Rename .spec files to .test files.
