+++ packages/address-watcher/package.json
@@ -5,7 +5,7 @@
   "private": true,
   "scripts": {
     "lint": "eslint .",
-    "test": "mocha -r ts-node/register src/**/*.spec.ts",
+    "test": "mocha -r ts-node/register src/**/*.test.ts",
     "build": "tsc",
     "server": "DEBUG=vulcanize:* node dist/server.js",
     "server:dev": "DEBUG=vulcanize:* nodemon --watch src src/server.ts",
