--- packages/address-watcher/src/test.spec.ts
@@ -1,3 +1,0 @@
-//
-// Copyright 2021 Vulcanize, Inc.
-//
