+++ packages/address-watcher/tsconfig.json
@@ -73,5 +73,5 @@
     "resolveJsonModule": true                       /* Enabling the option allows importing JSON, and validating the types in that JSON file. */
   },
   "include": ["src"],
-  "exclude": ["dist", "src/**/*.spec.ts"]
+  "exclude": ["dist", "src/**/*.test.ts"]
 }
