+++ packages/ipld-eth-client/package.json
@@ -6,7 +6,7 @@
   "main": "dist/index.js",
   "scripts": {
     "lint": "eslint .",
-    "test": "mocha -r ts-node/register src/**/*.spec.ts",
+    "test": "mocha -r ts-node/register src/**/*.test.ts",
     "build": "tsc"
   },
   "repository": {
