+++ packages/lighthouse-watcher/package.json
@@ -6,7 +6,7 @@
   "private": true,
   "scripts": {
     "lint": "eslint .",
-    "test": "mocha -r ts-node/register src/**/*.spec.ts",
+    "test": "mocha -r ts-node/register src/**/*.test.ts",
     "build": "tsc",
     "server": "DEBUG=vulcanize:* node dist/server.js",
     "server:dev": "DEBUG=vulcanize:* nodemon --watch src src/server.ts",
